[refactoring/daengle-57] Chat api 모듈, persistence 모듈 의존성 걷어내기
