+++ daengle-chat-api/src/main/java/ddog/chat/application/ChatService.java
@@ -1,7 +1,7 @@
 package ddog.chat.application;
 
 import ddog.domain.chat.ChatMessage;
-import ddog.persistence.dynamo.port.ChatMessagePersist;
+import ddog.domain.chat.port.ChatMessagePersist;
 import ddog.domain.chat.port.ChatRoomPersist;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
