+++ daengle-domain/src/main/java/ddog/domain/chat/port/ChatRoomPersist.java
@@ -3,9 +3,6 @@ package ddog.domain.chat.port;
 import ddog.domain.chat.ChatRoom;
 
 public interface ChatRoomPersist {
-
     ChatRoom enterChatRoom(Long userId, Long partnerId);
-
     void exitChatRoom(Long userId, Long partnerId);
-
 }
